+++ MoMoney.Core/Helpers/Constants.cs
@@ -18,7 +18,7 @@ public static class Constants
     public const int DEBIT_ID = 2;
     public const int CREDIT_ID = 3;
     public const int INCOME_ID = 4;
-    public const int EXPENSE_ID = 5; // anything >= 5 will be an expense
+    public const int EXPENSE_THRESHOLD = 5; // anything >= 5 will be an expense
 
     // month name array alias
     public static readonly string[] MONTHS = CultureInfo.CurrentCulture.DateTimeFormat.AbbreviatedMonthNames;