+++ MoMoney.Core/Helpers/Utilities.cs
@@ -36,11 +36,12 @@ public static class Utilities
         set { Preferences.Set("MaterialYou", value); }
     }
 
-    public static string ThemeMode
-    {
-        get { return Preferences.Get("ThemeMode", "System"); }
-        set { Preferences.Set("ThemeMode", value); }
-    }
+    // not working so just commenting it out for now
+    // public static string ThemeMode
+    // {
+    //     get { return Preferences.Get("ThemeMode", "System"); }
+    //     set { Preferences.Set("ThemeMode", value); }
+    // }
 
     public static Color GetColour(string lightColour, string darkColour) => UraniumUI.Resources.ColorResource.GetColor(lightColour, darkColour);
     public static Color GetColour(string colour) => UraniumUI.Resources.ColorResource.GetColor(colour);