+++ MoMoney.Core/ViewModels/Tabs/SettingsViewModel.cs
@@ -12,8 +12,7 @@ public partial class SettingsViewModel : ObservableObject
 
     [ObservableProperty] bool materialYouEnabled = Utilities.MaterialYouEnabled;
     [ObservableProperty] bool showSensitiveValuesEnabled = true;
-    [ObservableProperty]
-    public partial string SelectedThemeMode { get; set; }
+    //[ObservableProperty] public partial string SelectedThemeMode { get; set; }
 
     public List<ThemeInfo> Themes { get; } =
     [
@@ -28,7 +27,7 @@ public partial class SettingsViewModel : ObservableObject
     public SettingsViewModel(ILoggerService<SettingsViewModel> _logger)
     {
         logger = _logger;
-        SelectedThemeMode = Utilities.ThemeMode;
+        //SelectedThemeMode = Utilities.ThemeMode;
     }
 
     /// <summary>
@@ -119,26 +118,27 @@ public partial class SettingsViewModel : ObservableObject
         WeakReferenceMessenger.Default.Send(new UpdateHomePageMessage());
     }
 
-    /// <summary>
-    /// Saves selected theme mode to Preferences
-    /// </summary>
-    /// <param name="value"></param>
-    partial void OnSelectedThemeModeChanged(string value)
-    {
-        Utilities.ThemeMode = value;
-        switch (value)
-        {
-            case "Light":
-                Application.Current!.UserAppTheme = AppTheme.Light;
-                break;
-            case "Dark":
-                Application.Current!.UserAppTheme = AppTheme.Dark;
-                break;
-            default:
-                Application.Current!.UserAppTheme = AppTheme.Unspecified;
-                break;
-        }
-    }
+    // not working so just commenting it out for now
+    // /// <summary>
+    // /// Saves selected theme mode to Preferences
+    // /// </summary>
+    // /// <param name="value"></param>
+    // partial void OnSelectedThemeModeChanged(string value)
+    // {
+    //     Utilities.ThemeMode = value;
+    //     switch (value)
+    //     {
+    //         case "Light":
+    //             Application.Current!.UserAppTheme = AppTheme.Light;
+    //             break;
+    //         case "Dark":
+    //             Application.Current!.UserAppTheme = AppTheme.Dark;
+    //             break;
+    //         default:
+    //             Application.Current!.UserAppTheme = AppTheme.Unspecified;
+    //             break;
+    //     }
+    // }
 
     /// <summary>
     /// Goes to AdminPage.xaml.