+++ MoMoney/Views/MainPage.xaml.cs
@@ -1,3 +1,5 @@
+using CommunityToolkit.Maui.Behaviors;
+using CommunityToolkit.Maui.Core;
 using CommunityToolkit.Mvvm.Messaging;
 using MoMoney.Core.Helpers;
 
@@ -21,6 +23,35 @@ public partial class MainPage : ContentPage
 		WeakReferenceMessenger.Default.Register<ChangeTabMessage>(this, (r, m) =>
 		{
 			MainTabView.SelectedIndex = m.Value;
+		});
+
+		Application.Current!.RequestedThemeChanged += (s, e) => UpdateStatusBar();
+		UpdateStatusBar();
+	}
+
+	/// <summary>
+	/// Update Status bar colour to match requested theme
+	/// (Doesn't work in xaml, has to be here)
+	/// </summary>
+	void UpdateStatusBar()
+	{
+		var isDark = Application.Current!.UserAppTheme switch
+		{
+			AppTheme.Dark => true,
+			AppTheme.Light => false,
+			_ => Application.Current.RequestedTheme == AppTheme.Dark
+		};
+
+		var statusBarColor = Utilities.GetColour("secondaryContainer", "secondaryContainerDark");
+		var statusBarStyle = isDark
+			? StatusBarStyle.LightContent
+			: StatusBarStyle.DarkContent;
+
+		Behaviors.Clear();
+		Behaviors.Add(new StatusBarBehavior
+		{
+			StatusBarColor = statusBarColor,
+			StatusBarStyle = statusBarStyle
 		});
 	}
 }